requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.30
pymysql==1.1.0
cryptography==42.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.2.2
openpyxl==3.1.2
python-calamine==0.2.0
pdfplumber==0.11.0
python-dotenv==1.0.1
